Extract to offline_signing folder
