+++ base_layer/wallet/src/transaction_service/offline_signing/mod.rs
@@ -1,0 +1,2 @@
+pub mod models;
+pub mod offline_signer;
